UnitTest: 6AND, changes AND, testing the NOT.  (#9)
* UnitTest: circuit includes XNOR and XOR gates.

* UnitTest: circuit includes three NOT, FORK and one AND.

* UnitTest: 6AND, changes AND, changes NOT.

Changes in the number of inputs of the AND were tested. The works of 6AND, NOT has been cheked.

* UnitTest:8NAND, changes NAND.

Changes in the number of inputs of the NAND were tested. The works of 8NAND has been cheked.

* UnitTest:8NAND, changes NAND.

Changes in the number of inputs of the NAND were tested. The works of 6NAND, NOT has been cheked.

* Changed names of TestCases.